+++ lib/Controller/AmpacheController.php
@@ -331,11 +331,16 @@ class AmpacheController extends Controller {
 	 * @AmpacheAPI
 	 */
 	protected function get_indexes(string $type, ?string $filter, ?string $add, ?string $update, ?bool $include, int $limit, int $offset=0) : array {
-		if ($type === 'album_artist') {
+		if ($type === 'album_artist' || $type === 'song_artist') {
 			$type = 'artist';
 			list($addMin, $addMax, $updateMin, $updateMax) = self::parseTimeParameters($add, $update);
-			$entities = $this->artistBusinessLayer->findAllHavingAlbums(
-				$this->session->getUserId(), SortBy::Name, $limit, $offset, $filter, MatchMode::Substring, $addMin, $addMax, $updateMin, $updateMax);
+			if ($type === 'album_artist') {
+				$entities = $this->artistBusinessLayer->findAllHavingAlbums(
+					$this->session->getUserId(), SortBy::Name, $limit, $offset, $filter, MatchMode::Substring, $addMin, $addMax, $updateMin, $updateMax);
+			} else {
+				$entities = $this->artistBusinessLayer->findAllHavingTracks(
+					$this->session->getUserId(), SortBy::Name, $limit, $offset, $filter, MatchMode::Substring, $addMin, $addMax, $updateMin, $updateMax);
+			}
 		} else {
 			$businessLayer = $this->getBusinessLayer($type);
 			$entities = $this->findEntities($businessLayer, $filter, false, $limit, $offset, $add, $update);